+++ src/Zeus/Section/SectionException.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace GetOlympus\Zeus\Section;
+
+use GetOlympus\Zeus\Base\BaseException;
+
+/**
+ * Section exception.
+ *
+ * @package    OlympusZeusCore
+ * @subpackage Section
+ * @author     Achraf Chouk <achrafchouk@gmail.com>
+ * @since      2.1.8
+ *
+ */
+
+class SectionException extends BaseException
+{
+    //
+}